reorganized tests for use cases
