+++ epistemix_api/tests/test_job_routes.py
@@ -23,7 +23,6 @@ def client():
     
     # Clear and recreate tables for each test
     test_db_manager = get_database_manager(test_db_url)
-    test_db_manager.drop_tables()
     test_db_manager.create_tables()
     
     with app.test_client() as client:
